feat. Create Orden Sale

+++ src/main/java/com/rafaeltalavera/springboot/testwtldigital/models/entity/ItemSale.java
@@ -21,7 +21,7 @@ public class ItemSale implements Serializable {
 	private Long id;
 
 	@NotNull
-	private Double quantity;
+	private Integer quantity;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "product_id")
@@ -44,11 +44,11 @@ public class ItemSale implements Serializable {
 		this.id = id;
 	}
 
-	public Double getQuantity() {
+	public Integer getQuantity() {
 		return quantity;
 	}
 
-	public void setQuantity(Double quantity) {
+	public void setQuantity(Integer quantity) {
 		this.quantity = quantity;
 	}
